+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - remove direct dependency on `byteorder` crate #3031
 - make it possible to cancel message sending by removing the message #3034,
   this was previosuly removed in 1.71.0 #2939
+- synchronize Seen flags only on watched folders to speed up
+  folder scanning #3041
 
 ### Fixes
 - fix splitting off text from webxdc messages #3032